test=document_fix

+++ tools/ci_tools/ci_autoscan_arm.sh
@@ -1,4 +1,5 @@
-tart the CI task of unittest for op and pass.
+#!/bin/bash
+# Start the CI task of unittest for op and pass.
 set -x
 set -e
 
@@ -22,7 +23,7 @@ TARGET_LIST="ARM,OpenCL,Metal"
 #   WORKSPACE
 ####################################################################################################
 function auto_scan_test {
-  target_name=$1
+  local target_name=$1
 
   cd $WORKSPACE/lite/tests/unittest_py/rpc_service
   sh start_rpc_server.sh
@@ -47,7 +48,7 @@ function auto_scan_test {
 ####################################################################################################
 # Functions of compiling test.
 # Arguments:
-#   --target_list
+#   --target_list: can be ARM,OpenCL or ARM,Metal
 # Globals:
 #   WORKSPACE, PYTHON_VERSION
 ####################################################################################################
@@ -77,28 +78,34 @@ function compile_publish_inference_lib {
     fi
   done
 
+  if [[ "$build_opencl" == "ON" && "$build_metal" == "ON" ]]; then
+    echo "ERROR: OpenCL and Metal both turn on, which will cause unittests for Metal crash. You can only turn on one option."
+    exit 1
+  fi
+
   cd $WORKSPACE
 
   # Remove Compiling Cache
   rm -rf build.macos.*
 
-  # Step1. Compiling python installer on mac
-  cmd_line="./lite/tools/build_macos.sh --with_python=ON --with_opencl=$build_opencl --with_metal=$build_metal --with_arm82_fp16=ON --python_version=$PYTHON_VERSION arm64"
+  # Step1. Compiling python installer on mac M1
+  local cmd_line="./lite/tools/build_macos.sh --with_python=ON --with_opencl=$build_opencl --with_metal=$build_metal --with_arm82_fp16=ON --python_version=$PYTHON_VERSION arm64"
   $cmd_line
-  # Step2. Checking results: cplus and python inference lib.
-  build_dir=build.macos.armmacos.armv8.metal.opencl
 
-  if [ -d ${build_dir}/inference_lite_lib.armmacos.armv8.opencl.metal/python/install/dist ]; then
-    #install deps
-    python$PYTHON_VERSION -m pip install --force-reinstall  ${build_dir}/inference_lite_lib.armmacos.armv8.opencl.metal/python/install/dist/*.whl
-    python3.8 -m pip install -r ./lite/tests/unittest_py/requirements.txt
-  else
+  # Step2. Checking results: cplus and python inference lib
+  local whl_path=$(find ./build.macos.armmacos.armv8.* -name *whl)
+  if [[ -z "$whl_path" ]]; then
     # Error message.
     echo "**************************************************************************************"
+    echo -e "$whl_path not found!"
     echo -e "Compiling task failed on the following instruction:\n $cmd_line"
     echo "**************************************************************************************"
     exit 1
   fi
+
+  # Step3. Install whl and its depends
+  python$PYTHON_VERSION -m pip install --force-reinstall $whl_path
+  python3.8 -m pip install -r ./lite/tests/unittest_py/requirements.txt
 }
 
 function run_test() {
@@ -108,6 +115,18 @@ function run_test() {
   for target in ${targets[@]}; do
     auto_scan_test $target
   done
+}
+
+function pipline() {
+  # Compile
+  compile_publish_inference_lib --target_list=$1
+
+  # Run unittests
+  run_test $1
+
+  # Uninstall paddlelite
+  python$PYTHON_VERSION -m pip uninstall -y paddlelite
+  echo "Success for targets:" $1
 }
 
 function main() {
@@ -125,15 +144,16 @@ function main() {
     esac
   done
 
-  # Compile
-  compile_publish_inference_lib --target_list=$TARGET_LIST
+  local targets=(${TARGET_LIST//,/ })
+  for target in ${targets[@]}; do
+    if [[ "$target" == "OpenCL" ]]; then
+      pipline "ARM,OpenCL"
+    elif [[ "$target" == "Metal" ]]; then
+      pipline "Metal"
+    fi
+  done
 
-  # Run unittests
-  run_test $TARGET_LIST
-
-  # Uninstall paddlelite
-  python$PYTHON_VERSION -m pip uninstall -y paddlelite
-  echo "Success."
+  echo "Success for targets:" $TARGET_LIST
 }
 
 main $@
